+++ src/main/java/com/estifie/expensetracker/service/SubscriptionServiceImpl.java
@@ -2,6 +2,8 @@ package com.estifie.expensetracker.service;
 
 import com.estifie.expensetracker.dto.expense.ExpenseCreateDTO;
 import com.estifie.expensetracker.dto.subscription.SubscriptionCreateDTO;
+import com.estifie.expensetracker.enums.Permission;
+import com.estifie.expensetracker.exception.subscription.SubscriptionNotFoundException;
 import com.estifie.expensetracker.exception.user.UserNotFoundException;
 import com.estifie.expensetracker.model.Subscription;
 import com.estifie.expensetracker.model.User;
@@ -9,11 +11,14 @@ import com.estifie.expensetracker.repository.SubscriptionRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
+@Service
 public class SubscriptionServiceImpl implements SubscriptionService {
     private final SubscriptionRepository subscriptionRepository;
     private final ExpenseService expenseService;
@@ -37,7 +42,8 @@ public class SubscriptionServiceImpl implements SubscriptionService {
 
     public void deactivate(String id) {
         Optional<Subscription> optionalSubscription = subscriptionRepository.findById(id);
-        if (optionalSubscription.isEmpty()) return;
+        if (optionalSubscription.isEmpty())
+            throw new SubscriptionNotFoundException();
 
         Subscription subscription = optionalSubscription.get();
         subscription.setActive(false);
@@ -45,8 +51,20 @@ public class SubscriptionServiceImpl implements SubscriptionService {
         subscriptionRepository.save(subscription);
     }
 
-    public void delete(String id) {
-        subscriptionRepository.deleteById(id);
+    public void delete(String id, boolean hardDelete) {
+        Subscription subscription = subscriptionRepository.findById(id)
+                .orElseThrow(SubscriptionNotFoundException::new);
+
+        subscription.setDeletedAt(LocalDateTime.now());
+
+        boolean canHardDelete = userService.hasPermission(subscription.getUser()
+                .getUsername(), Permission.HARD_DELETE_SUBSCRIPTION.name());
+
+        if (canHardDelete && hardDelete) {
+            subscriptionRepository.deleteById(id);
+        } else {
+            subscriptionRepository.save(subscription);
+        }
     }
 
     public Optional<Subscription> findById(String id, boolean fetchDeleted) {
@@ -63,7 +81,7 @@ public class SubscriptionServiceImpl implements SubscriptionService {
 
     @Scheduled(cron = "0 0 0 * * *")
     private void processRecurringPayments() {
-        List<Subscription> subscriptions = subscriptionRepository.findAllByNextPaymentDate(LocalDate.now());
+        List<Subscription> subscriptions = subscriptionRepository.findAllByNextBillingDate(LocalDate.now());
 
         for (Subscription subscription : subscriptions) {
             processSubscription(subscription);
